added add_random_bits_between_bytes and converted from long double to double

+++ GaloisFieldHillCipher/GaloisFieldHillCipher.cpp
@@ -5,7 +5,7 @@
 
 int main()
 {
-	long double** out_matrix = NULL;
+	double** out_matrix = NULL;
 
 	STATUS_CODE return_code = generate_encryption_matrix(&out_matrix, 4, DEFAULT_PRIME_GALOIS_FIELD);
 	
@@ -13,7 +13,7 @@ int main()
 	{
 		for (uint32_t column = 0; column < 4; ++column)
 		{
-			printf("%lf ", out_matrix[row][column]);
+			printf("%f ", out_matrix[row][column]);
 		}
 		printf("\n");
 	}
